idea提交

+++ src/main/java/co/yiiu/pybbs/controller/admin/TagAdminController.java
@@ -52,11 +52,8 @@ public class TagAdminController extends BaseAdminController {
   @RequiresPermissions("tag:add")
   @PostMapping("/add")
   public String add(String name, String description, Integer topicCount) {
-
-//    System.out.println(adminUser.getUsername());
     tag.setName(name);
     tag.setDescription(description);
-    //tag.setTopicCount(topicCount);
     tag.setAdminId(null);
     tagService.insertTag(tag);
     return redirect("/admin/tag/list");
@@ -89,6 +86,8 @@ public class TagAdminController extends BaseAdminController {
   public Result delete(Integer id) {
     //Tag tag = tagService.selectById(id);
     //if (tag.getTopicCount() > 0) return error("标签还关联着话题，要先把相关联的话题都删了，这个标签才能删除");
+    Tag tag = tagService.selectById(id);
+    if (tag.getTopicCount() > 0) return error("板块还关联着话题，要先把相关联的话题都删了，这个板块才能删除");
     tagService.delete(id);
     return success();
   }
